pyproject.toml
[build-system]
# Minimum requirements for the build system to execute.
requires = ["setuptools", "wheel"]  # PEP 508 specifications.
build-backend = "setuptools.build_meta"

[project]
name = "populse_mia"
#version = "2.5.3_dev"
dynamic = ["version"]
authors = [
    {name = "Populse team", email = "populse-support@univ-grenoble-alpes.fr"},
]
description = "populse_mia"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "CeCILL-B"}
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "capsul >= 2.5.0, < 3.0.0",
    "cryptography",
    "matplotlib",
    "mia-processes >= 2.5.0, < 3.0.0",
    "nibabel",
    "nipype",
    "pillow",
    "populse-db >= 2.5.0, < 3.0.0",
    "pre-commit",
    "pyqt5",
    "python-dateutil",
    "pyyaml",
    "scikit-image",
    "scipy",
    "snakeviz",
    "soma-base >= 5.2.0, < 6.0.0",
    "soma-workflow >= 3.2.2",
    "six >= 1.13",
    "traits",
]

[project.optional-dependencies]
doc = [
    "sphinx >=1.0",
]

[project.urls]
homepage = "http://populse.github.io/populse_mia/"
repository = "https://github.com/populse/populse_mia"

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
atomic = true
profile = "black"
line_length = 79
skip_gitignore = true
#known_first_party = ['populse_mia']

[tool.coverage.run]
omit = [
    "populse_mia/user_interface/data_viewer/anatomist_2/*",
    "populse_mia/user_interface/data_viewer/anatomist/*",
    "processes/*",
    "/tmp/*"
]

[tool.interrogate]
ignore-init-method = true
ignore-init-module = false
ignore-magic = false
ignore-semiprivate = false
ignore-private = false
ignore-property-decorators = false
ignore-module = false
ignore-nested-functions = false
ignore-nested-classes = true
ignore-setters = false
fail-under = 100
exclude = ["setup.py", "docs", "build", "processes"]
verbose = 2
quiet = false
color = true
omit-covered-files = false
